requirements.txt
anthropic>=0.40.0
chromadb>=0.5.0
yt-dlp>=2024.11.0
youtube-transcript-api>=1.0.0
pymupdf>=1.24.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
tiktoken>=0.8.0
rich>=13.9.0
httpx>=0.27.0
tqdm>=4.66.0
python-multipart>=0.0.9
